requirements.txt

transformers
opencv-python
pytesseract
torch
torchvision
Pillow
streamlit
numpy
